Refactor quarter state to its own file
